+++ pydidas/gui/mixins/view_results_mixin.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -19,11 +21,12 @@ WorkflowTree results when running the pydidas WorkflowTree.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-2022, Malte Storm, Helmholtz-Zentrum Hereon"
-__license__ = "GPL-3.0"
+__copyright__ = "Copyright 2021-, Helmholtz-Zentrum Hereon"
+__license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Development"
 __all__ = ["ViewResultsMixin"]
+
 
 import os
 
@@ -76,8 +79,6 @@ class ViewResultsMixin:
                 "source_hash": self._RESULTS.source_hash,
             }
         )
-        self._data_axlabels = ["", ""]
-        self._data_axunits = ["", ""]
         self.connect_view_results_mixin_signals()
         self._update_choices_of_selected_results()
         self.__export_dialog = PydidasFileDialog(
@@ -128,8 +129,7 @@ class ViewResultsMixin:
         Clear all curves / images from the plot and disable any new updates.
         """
         self._config["plot_active"] = False
-        for _plot in [self._widgets["plot1d"], self._widgets["plot2d"]]:
-            _plot.clear_plot()
+        self._widgets["plot"].clear_plots()
 
     @QtCore.Slot(bool, object, int, object, str)
     def update_result_selection(
@@ -173,7 +173,6 @@ class ViewResultsMixin:
         """
         if not self._config["plot_active"]:
             return
-        _dim = 1 if self._config["plot_type"] in ["1D plot", "group of 1D plots"] else 2
         _node = self._config["active_node"]
         _data = self._RESULTS.get_result_subset(
             _node,
@@ -181,18 +180,12 @@ class ViewResultsMixin:
             flattened_scan_dim=self._config["data_use_timeline"],
             squeeze=True,
         )
-        self._data_axlabels = _data.axis_labels.copy()
-        self._data_axunits = _data.axis_units.copy()
         if self._config["plot_type"] == "group of 1D plots":
-            self._widgets["plot_stack"].setCurrentIndex(0)
             self._plot_group_of_curves(_data)
         elif self._config["plot_type"] == "1D plot":
-            self._widgets["plot_stack"].setCurrentIndex(0)
             self._plot1d(_data, replace=True)
         elif self._config["plot_type"] in ["2D full axes", "2D data subset"]:
-            self._widgets["plot_stack"].setCurrentIndex(1)
             self._plot_2d(_data)
-        self._widgets[f"plot{_dim}d"].setGraphTitle(self._RESULTS.result_titles[_node])
 
     def _plot_group_of_curves(self, data):
         """
@@ -244,7 +237,7 @@ class ViewResultsMixin:
             )
         if not isinstance(data.axis_ranges[0], np.ndarray):
             data.update_axis_ranges(0, np.arange(data.size))
-        self._widgets["plot1d"].plot_pydidas_dataset(
+        self._widgets["plot"].plot_data(
             data,
             replace=replace,
             legend=legend,
@@ -266,7 +259,7 @@ class ViewResultsMixin:
         _dim0, _dim1 = self._config["active_dims"]
         if _dim0 > _dim1:
             data = data.transpose()
-        self._widgets["plot2d"].plot_pydidas_dataset(
+        self._widgets["plot"].plot_data(
             data, title=self._RESULTS.result_titles[self._config["active_node"]]
         )