+++ pydidas/widgets/silx_plot/pydidas_plot1d.py
@@ -18,11 +18,14 @@ Module with PydidasPlot1D class which adds configurations to the base silx Plot1
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-2022, Malte Storm, Helmholtz-Zentrum Hereon"
-__license__ = "GPL-3.0"
+__copyright__ = "Copyright 2021-, Helmholtz-Zentrum Hereon"
+__license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Development"
 __all__ = ["PydidasPlot1D"]
+
+
+import inspect
 
 from qtpy import QtCore
 from silx.gui.plot import Plot1D
@@ -35,13 +38,18 @@ class PydidasPlot1D(Plot1D):
 
     def __init__(self, parent=None, backend=None):
         Plot1D.__init__(self, parent, backend)
+        self._allowed_kwargs = [
+            _key
+            for _key, _value in inspect.signature(self.addCurve).parameters.items()
+            if _value.default is not inspect.Parameter.empty
+        ]
 
         self.getRoiAction().setVisible(False)
         self.getFitAction().setVisible(False)
 
         self.setAttribute(QtCore.Qt.WA_DeleteOnClose)
 
-    def plot_pydidas_dataset(self, data, replace=True, title=None, legend=None):
+    def plot_pydidas_dataset(self, data, **kwargs):
         """
         Plot a pydidas dataset.
 
@@ -49,35 +57,30 @@ class PydidasPlot1D(Plot1D):
         ----------
         data : pydidas.core.Dataset
             The data to be plotted.
-        replace : bool
-            Keyword to replace the active plot. If False, the new graph will be added
-            to the existing graph. THe default is True.
-        title : Union[None, str], optional
-            The title for the plot. If None, no title will be added to the plot.
-        legend : Union[None, str], optional
-            If desired, a legend entry for this curve. If None, no legend
-            entry will be added. The default is None.
+        **kwargs : dict
+            Additional keyword arguments to be passed to the silx plot method.
         """
-        if replace:
+        if kwargs.get("replace", True):
             self.clear_plot()
 
-        _x_axlabel = data.axis_labels[0] + (
-            " / " + data.axis_units[0] if len(data.axis_units[0]) > 0 else ""
+        self.setGraphXLabel(
+            data.axis_labels[0]
+            + (" / " + data.axis_units[0] if len(data.axis_units[0]) > 0 else "")
         )
-        _y_axlabel = data.data_label + (
-            " / " + data.data_unit if len(data.data_unit) > 0 else ""
+        self.setGraphYLabel(
+            data.data_label
+            + (" / " + data.data_unit if len(data.data_unit) > 0 else "")
         )
         self.addCurve(
             data.axis_ranges[0],
             data.array,
             linewidth=1.5,
-            legend=legend,
-            replace=replace,
+            **{
+                _key: _val
+                for _key, _val in kwargs.items()
+                if _key in self._allowed_kwargs
+            },
         )
-        self.setGraphYLabel(_y_axlabel)
-        self.setGraphXLabel(_x_axlabel)
-        if title is not None:
-            self.setGraphTitle(title)
 
     def clear_plot(self):
         """